#12 [REFACTOR] : UI 디테일 수정
