+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
         </activity>
         <activity
             android:name=".feature.MainActivity"
-            android:exported="true"
+            android:exported="false"
             android:screenOrientation="portrait"></activity>
         <activity
             android:name=".feature.example.xml.ExampleActivity"
