+++ app/src/main/java/com/univoice/data_local/UserPreferencesDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.univoice.data_remote.datasourceimpl
+package com.univoice.data_local
 
 import android.content.Context
 import androidx.datastore.preferences.core.edit
