+++ app/src/main/res/drawable/sel_login_button.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="@drawable/shape_mint_fill_40_rect" android:state_enabled="true" />
+    <item android:drawable="@drawable/shape_mint400_fill_40_rect" android:state_enabled="true" />
     <item android:drawable="@drawable/shape_gray200_fill_40_rect" android:state_enabled="false" />
 </selector>
