+++ app/src/main/res/drawable/shape_white_line_40_rect.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
-    <solid android:color="@color/font_W01" />
+    <solid android:color="@color/white" />
     <stroke
         android:width="1dp"
         android:color="@color/mint_600" />
